+++ MM-control-01/motion.h
@@ -9,7 +9,7 @@
 
 extern int8_t filament_type[EXTRUDERS];
 extern const int filament_lookup_table[8][3]; // [X][Y]Two-dimensional Array of extruder and used variables
-extern const int IDLER_PARKING_STEPS;
+extern const uint8_t IDLER_PARKING_STEPS;
 extern uint16_t BOWDEN_LENGTH;
 
 void home(bool doToolSync = false);
